requirements.txt
pyperclip==1.8.2
py_notifier==0.3.1
ffmpeg_python==0.2.0
tqdm==4.60.0
requests==2.25.1
ffmpeg==1.4
PyYAML==5.4.1
urllib3==1.25.11
